+++ runtime-discovery/dockingBike/src/main/java/dockingBike/DockingRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DockingRepository extends JpaRepository<Docking, Long> {
+public interface DockingRepository extends JpaRepository<Docking, Integer> {
 
-	Docking findByDockingId(Long dockingId);
+	Docking findByDockingId(int dockingId);
 
 }
